// src/schemas/product.ts
export interface Product {
    id: string;
    title: string;
    description: string;
    price: number;
    discountedPrice: number;
    image: ProductImage;
    rating: number;
    tags: string[];
    reviews: Review[];
}

export interface Review {
    id: string;
    username: string;
    rating: number;
    description: string;
}

export interface ProductImage {
    url: string;
    alt: string;
}